Django_3/Final/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
cffi==1.17.1
channels==4.2.2
constantly==23.10.4
cryptography==45.0.5
csvfile==3.4.0
daphne==4.2.1
Django==4.2.23
django-bootstrap-v5==1.0.11
django-jquery==3.1.0
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.8.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dotenv==1.1.1
service-identity==24.2.0
setuptools==80.9.0
soupsieve==2.7
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
zope.interface==7.2
